fix wrong size and sign of speed when reading encoder_speed values

+++ roboclaw/include/roboclaw/io/read_commands/encoder_speed.hpp
@@ -13,7 +13,7 @@ namespace read_commands
 {
 struct encoder_speed_values
 {
-    uint16_t speed;
+    int32_t speed;
     bool forward;
 };
